Added some funcionalities for buying and reservation ticket, exceptions and start with testing

+++ TicketService/src/main/java/com/ftn/service/ReservationService.java
@@ -2,9 +2,11 @@ package com.ftn.service;
 
 
 import com.ftn.dtos.ReservationDto;
+import com.ftn.exceptions.AplicationException;
 import com.ftn.exceptions.EntityNotFoundException;
 import com.ftn.model.Reservation;
 import com.ftn.model.Ticket;
+import com.ftn.model.User;
 import com.ftn.repository.ReservationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -43,7 +45,20 @@ public class ReservationService {
     }
 
     public void deleteReservation(Long id){
-        reservationRepository.deleteById(id);
+
+        User u = userService.getloggedInUser();
+
+        List<Reservation> reservationsOfUser = new ArrayList<>();
+        reservationsOfUser.addAll(u.getReservations());
+
+            Reservation r = findOneReservation(id);
+
+            if(reservationsOfUser.contains(r)) {
+                ticketService.deleteTicket(r.getTicket().getId());
+                reservationRepository.deleteById(id);
+            }else{
+                throw new AplicationException("Can't cancel other users reservations!");
+            }
     }
 
     public void deleteAll(){
@@ -52,6 +67,25 @@ public class ReservationService {
 
     public Boolean ifExist(Long id){
         return reservationRepository.existsById(id);
+    }
+
+    public List<ReservationDto> reservationOfUser(){
+
+        List<Reservation> reservations = new ArrayList<>();
+        List<ReservationDto> reservationDtos = new ArrayList<>();
+
+        User u = userService.getloggedInUser();
+
+        if (u == null) {
+            throw new AplicationException("You must be logged in to get your reservations.");
+        }
+
+        reservations.addAll(u.getReservations());
+        for (Reservation r : reservations) {
+            reservationDtos.add(mapToDTO(r));
+        }
+        return reservationDtos;
+
     }
 
     public ReservationDto mapToDTO(Reservation reservation){
